Add Survey.most_recent_for_each_title.
When you re-import a survey with the same title as an existing one,
Surveyor appends a serial number to the title. This method strips
those serial numbers off and returns the most recently created survey
for each title.

This commit also applies the survey load speedup from c7715c1 to the
Survey model specs.