Load records from the warehouse in blocks. #1642.
This loads everything in blocks of 2500 except for the participant
state-influencing events, instruments, and link_contacts. They need to be
clustered in memory, so they'd have to be blocked by participant which would
be pretty complicated. Practically speaking, there are only 10k of them
or so in the GCSC data, so they won't have too much memory impact.