Handle multiple question variables in the simplest possible way. #1645.
Theoretically it should be possible to evaluated the dependencies and only set
the value that was set. However, that would be require multiple passes (to
make sure that the field on which the question is dependent had already been
imported) and still might fail. With question templating we should be able to
eliminate multiple-question variables (since they are mostly just wording
differences), this stopgap should suffice for now.