Handle state-affecting events, etc., by blocks as well. #1642.
The blocks are of event IDs, so the actual number of records retrieved for
dependent records may be more or less than the block size. Also, since the
all the events for a particular participant need to be handled in the same
block, the event ID lists may be slightly longer than the block size.