Configure the instrument transformer to use a subprocess. #1646.
Since the enumerator is very dependent on the Surveyor API, and the Surveyor
API won't work outside of a full-up Rails app, the instrument enumerator
needs to run with the core rails app environment loaded. Doing that within
the ETL ruby process directly would be a huge mess, so do it in a separate
process using `rails runner`.