DataMapper objects don't have an automatic "id". #1642.
This is a difference from AR, where the objects always have an `id` attribute
even if the PK is called something else. In DM the corresponding attribute is
called `key` and is an array (since DM supports CPKs by default). It's easier
just to use the actual PK name in this instance.