Use Rack::Test for testing APIs.  #1804.
Capybara is too high-level an abstraction for API testing: for example,
it offers very limited control over choice of HTTP method.  It is
possible to do all of this without metaclass tomfoolery by setting the
Capybara session driver to :rack_test, but I can't think of a good
reason to deal with all of that extra machinery.