Add spec for acceptable multi-segment events. #1817.
The importer, given an event that doesn't correspond to an existing scheduled
segment, need to either be able to uniquely determine which segment to
schedule or safely defer the event until all other events are scheduled. This
spec encapsulates what that requires from the structure of the template.

It also reveals that the birth visit will have to be special-cased in the
importer, since it unavoidably occurs alone on two separate segments.