Allow for checking for a schedule before registration. #1817.
If someone attempts to get the schedule or any of the schedule-derived data
before the subject is registered, PSC will 404. This module will return an
empty collection in this case. This is reasonable because there's no way for
a legitimate schedule to ever be completely empty and it allows for caching
of this negative result.