/people/1/start_instrument -> /contact_links/1/start_instrument.  #1924.
This commit also derives a Person from the identified ContactLink and
collapses the utility methods used by start_instrument into the action.
I plan to move a lot of this out of the controller, and I find that for
code of this size, having it all in one method helps me see what's going
on.