Unify cucumber configurations.
The cucumber executable only looks for step definitions in directories below
the directory containing the feature you are executing, unless you specify
others. Thus, the way Core has its features divided up requires `-r features`
in order for you to be able to run one feature at a time.

This change adds the appropriate switch, so long as you run `cucumber` from
the root of the project. It doesn't seem possible to make the path absolute
and so remove the WD restriction.