Contact, event, and instrument merge.  #1978.
A note on the adapters: I'm using the _before_type_cast variants of the
ActiveRecord attribute methods because it simplifies testing.

For example: Contact#distance_traveled and Event#incentive_cash are
represented in Core as decimal fields, but they are reported from the
client as strings.  As such, it is possible to end up with a situation
where this occurs:

    c.distance_traveled = str
    c.distance_traveled.to_s == str   # => false

due to BigDecimal interpreting the string as a decimal number.

Getting around this requires either special-casing value generation or
working with the attributes in their uncasted form.  I really don't
think special-casing value generation is necessary -- so long as values
can be compared for equality, the nature of those values is irrelevant
to the merge process -- and I like the symmetry of roundtripping.