Crude query profiler based on AS:Notifications.
Set PROFILE_DB=true to get a list of three top-ten list of queries run during
the tests:

* Most commonly executed
* Longest total time spent
* Highest average time

Also shows the fraction of the time spent in the database vs. the whole
execution. Assuming I did this correctly, you can see that even in the model
specs the actual database time is a small fraction of the time spent. However,
lots of the most frequent calls are INSERTs, so there is probably some chunk
of time spent in model instance creation and validation (via FactoryGirl).