Customizable IDs for entities in the fieldwork fixture.  #1975.
UUIDs of entities in merge integration tests are generated when
preconditions are set, and there's no way for Core's current sync API to
accept precalculated UUIDs from a field client.  Hence, this bit of
database twiddling.

Gross, yes, but this is the best I've come up with that doesn't require
API changes.  If we can afford the luxury of such changes in the future,
maybe it'll be worth it to revisit this smell then.