Associate contacts with their persons.  #2047, #1975.
I think that this can be done without opening up the hierarchy structure
like this (i.e. just read the ID out of the contact object), and since
hierarchy tracking is internal to the merge process, it doesn't really
matter how it's done.  The upshot: this stuff might be removed in a
future commit.