Use adapted objects for representing ancestry.  #2047.
I don't have a justification for this yet; however, doing this doesn't
take any more memory than referencing hashes, and all Superposition
behaviors are still satisfied by this change.

I also think that because the rest of the merge codebase deals with
adapters, the ancestor data should be the same way.  We'll see if it
makes a difference.