Splat args to redis.del. Done with #2062.
For no reason I can determine, redis-rb's del function will not work with an array of keys
when connected to the memory-only, CI-specific redis. It will only accept the keys as
separate args; hence the *. The test for this block of code works fine with either the
splat or an array when connected to my homebrew-installed redis, however.