Rely less on op enumerator for importer test setup. Done with #2062.
`create_warehouse_record_via_core` is nice for the bulk importer tests so that we can rely
on the FactoryGirl factories to fill in valid values for the test data. However, since
#2114 events and related tables have had filtering applied in the op enumerator. Working
around this filtering by dynamically adding values to the factory-created instances or by
adding more factories is more trouble than it is worth.

This commit switches the event-and-related import tests to use the warehouse API directly
to add test import source records.