Move system configuration accessors to lib. Closes #2269.
This makes it possible to use the configuration accessors from other initializers
without messing around with initializer order. Converting the accessors from module-level
methods to instance methods on a class also makes it easier to test them; this commit
includes tests for all the existing accessors.

While the implementations are all now on NcsNavigator::Core::Configuration, the
NcsNavigatorCore module has delegators for all of them pointing to a single global
instance. This both avoids needing to make a lot of small changes all over the code in
a bugfix release and gives them somewhat shorter names.