Ensure that the target can be overwritten. #2266.
The configuration images as deployed by chef are read-only. This is reasonable
enough, but they carry their permissions with them when they are copied into
the app. This means that the second deployment that attempts to copy them
fails with a permission error. This change prevents that error by making the
target asset writable just in time.