Copy ancestry when resolving response groups.
When a ResponseGroup is saved, the responses in that group are assigned
response set IDs from ancestor data.  There exists a quirk in
Superposition that prevents efficient retrieval of ancestors for current
(i.e. in-Core) state, so Superposition does not currently set ancestors
for current state.

While gross, that compromise currently works fine for everything but
Responses.  (Responses are the only entity that make use of that
ancestry data.)

This commit piles on another hack to copy ancestry from one group to
another, which solves the ancestry problem for Responses.  In the
future, a cleaner solution must be found.