Replicate entity merge examples for QRSes.  #2291.
At some point, it _might_ be a good idea to make this the same example
group used by the operational data entities; however, the setup for
the tests is quite different.

QuestionResponseSet#changed? intentionally just returns false for now,
because that's all that's needed to get these tests passing.  I'll
figure out a useful implementation later.