qrs.patch(qrs) should not change qrs.  #2291.
This is important because there are several merge examples in which the
current set is identical to another set (say, proposed).  For
simplicity, no diff code was written into the merge code; the code is
written to always execute #patch on current with some other set.

I still think this is a good idea, but it does mean that e.patch(e) must
not change the value of e.  Merge-by-attribute on ActiveRecord objects
accomplishes this trivially; patching QuestionResponseSets requires a
little more work.