Handwritten adapter classes.  #2291.
The motivation for this (quite large) change is that Response models
need to have a response set ID to pass database validation.  However,
response set IDs are not explicitly set by the field client -- they must
be implied by the structure of the JSON object.

Therefore, we cannot just copy the data for the response object into a
Response model.  We do, however, have access to ancestor data for all
entities, and the response set UUID is available from the response.

The simplest way to remedy this (IMO) is to make the
ResponseHashAdapter's #to_model write in the response set ID, but doing
that isn't something that's worth writing a generator for.  Hence, this.