Stub out Psc::ScheduledActivity.
We currently have (at least!) three implementations of this concept in
Cases.  That's two too many.

Another, more important motivation: The fieldwork merge process must
communicate with PSC to close out activities completed in the field.  It
doesn't make sense to implement that until we solidify the PSC
foundation in Cases.