Include participant public IDs in ResponseSet JSON.  #2335.
Yes, this does mean even more fun user torture.  I guess something like

    default_scope :include => :participant

or whatever ActiveRecord calls it now might be a quick fix, but it's not
the right thing to do, because it looks extremely awkward and it incurs
a cost even when the participant's attributes aren't needed.

The real fix is to get Surveyor to play nice with eager loading all the
way from e.g. Survey#most_recent_for_access_code, but that's a bit
further down the line.