Field::Merge#save needs to return merged models.  #2302.
If we're going to update the state of those models in PSC, we need to
know which models were affected.  The most straightforward way to get
that information IMO is from the save process.

This commit also modifies QuestionResponseSet to return a public ID
(even though it doesn't persist such an ID); making all mergeable
entities respond to #public_id means that I don't have to keep passing
said ID across a bunch of methods.