Eliminate adapter-model equality.  #2302.
Treating a fieldwork model adapter equal to the model itself has been
useful in the past, but it's not useful anymore.

Why?

Superposition#current{events, instruments, participants} should return
actual model objects, because they'll be used in processes that retrieve
and update PSC scheduled activities.  Said processes expect their
objects to respond to #sa_activity_ids and other bits that really exist
on the model objects.

It's possible to delegate those messages via the adapters, but that just
makes the adapters a mess.

Hence the change in this commit.  We can still have our one-line
expectations; we just need a different matcher.