Build InstrumentPlans in Psc::ScheduledActivity.  #2426.
As part of this commit, the construction of implied entities was moved
to Psc::ScheduledActivity.

Why?

It's much easier to build plans when you have access to the activity and
everything it implies.

In order to build an InstrumentPlan, you need the instrument and surveys
implied by the activity; and if you're going to work on the instrument,
the easiest way to check if two instruments are the same is to check
their structure, which includes checking the event and person referenced
by that instrument.

One *could* do all of this by just checking labels, but
Psc::ScheduledActivity is IMO much nicer to work with.

Next up: modify Field::ScheduledActivityReport and
Psc::ScheduledActivityReport::EntityResolution to work with
InstrumentPlans.