Allow the GEMSET to be set from outside.
There are two separate CI builds for Cases. They are simultaneously
kicked off on the same CI server whenever there's a push. Most of
the time, using the same gemset for each is fine since the gems are
locked with Gemfile.lock. However, if a push happens in the midnight
to 5am window in which the gemset is purged before each CI build,
the two jobs will attempt to purge simultaneously. This will cause
on of them to fail.

Allowing GEMSET to be provided externally will allow the jobs to use
separate gemsets, avoiding this problem.