Don't store model resolution results in implications.  #2426.
There's no need to ensure two implication objects point to the same
object in memory.  This lets us eliminate
ScheduledActivityReport::Collection; additionally, we longer abuse the
equality notions of Struct.

Also, it's less code:

    6 files changed, 147 insertions(+), 218 deletions(-)

This became important when #derive_implied_entities was moved from
Psc::ScheduledActivityReport to Psc::ScheduledActivity, with the
implication (ha) that implication objects could now be instantiated
multiple times.  Keeping around the hidden requirement for object
equality was just too much of a pain to deal with.