Store implied surveys in InstrumentPlans.  #2426.
This better fits what a Psc::InstrumentPlan is supposed to provide.  It
also makes it possible to test participant type and order without having
to instantiate Psc::ScheduledActivity instances.

To facilitate this change, a few changes to Psc::SA::Implications
objects were made:

1. participant_type and order attributes were moved from Instrument to
   Survey.
2. SurveyReference, a struct with (currently) one member, was
   introduced.

(1) was done because participant_type and label are really attributes of
the _survey_, not the operational data for the instrument.  (2) is a
consequence of (1): the additional attributes on Survey means that it is
no longer possible to construct a Survey from a single references label,
so we instead store the referenced code and use that for lookup.