Link instruments -> plans in fieldwork.  #2426.
A note on the Fingerprint module, which implements content addressing
for Instruments (and most other implied entities as a result):

What I'd *really* like is to use activity names for instrument plan IDs, so
that an instrument named "Birth Interview" references an instrument plan
with ID "Birth Interview".

Unfortunately, I don't think that's possible.  AFAICT, there is nothing
preventing two activities that have identical names but different
instrument references.  If we were to use the instrument's activity
name, we would be using an unfounded assumption that could lead to some
odd errors down the line: a plan for one instrument showing up for
another instrument, for example.

Should that assumption be proven correct later on, I'd be happy to get
rid of Fingerprint.  Until then, though, this is the safer option.