Field::Superposition now builds QuestionResponseSets.
This change isn't a prerequisite of anything, but I think it's cleaner.
A QRS is a merge contrivance, but it uses the same data structure as all
other entities in a superposition.

Doing this also demonstrates the use for a Field::Superposition#build
method that can be used to properly construct a superposition in one
call.