Additional accessors for ResponseSet merge adapters.  #2590.
Note: instrument_id is not an attribute that we receive from Field, but
that's okay.  Under normal circumstances, this means that original and
proposed values for instrument_id will be nil, and the current value
will be inferred by structure in ResponseSetHashAdapter#to_model.  We
therefore have the case [nil, ~nil, nil], which the merge algorithm can
handle as a non-conflicting case.

The schema validation step doesn't protect us from injection of
instrument public IDs, but this is okay too.  It probably should, but
it's not yet feasible to lock down Surveyor's response set serialization
-- if we want to do that, we may have to customize it for Cases rather
than relying on Surveyor's default serialization strategy.

Anyway.  Cases should never inject instrument IDs into response sets.
If it does, it's an error and must be fixed.  If Field sends us
instrument IDs, then we have two possibilities:

1. The instrument ID is the instrument ID we'd normally derive
2. The instrument ID is not the instrument ID we'd normally derive

In case 1, we have [nil, A, A] and can handle that as a non-conflicting
case.  In case 2, we have [nil, A, B], which is treated as a conflict;
we record this.  Finally, the instrument ID will be taken from hierarchy
anyway, so there is no danger of injected attributes messing up
relationships.