Display merge log on scenario failure.  #2590.
The weird logdev.rewind bit in Merge is needed to conform to IO's
interface.  (#string isn't in that interface, because it doesn't make
sense for a lot of IO types.)

We want to try to get all log data in the merge object, so we rewind
(and since some IOs aren't rewindable, ignoring any errors), and read
what we can.

The Cucumber tests can get away with using #string because we know that
@logdev will respond to it.