Hack tests into the green.  #2590.
I'm not really sure what a better way to fix this is.  Every adapted
response set *should* have an instrument ancestor, but the tests haven't
yet been revised to generate object hierarchies.  They *could* be, but
there's other things (viz. #2525) that need to get done.

For now, then, this is what I'm going with.  Yeah, it's cheesy, but it
works.  It *will* cause false positives in the merge (i.e. nil
instrument ID), but that's something that needs to be remedied in
ResponseSet.