Eliminate all existing response sets for a test survey.  #2618.
When multiple response sets for a survey are present, Cases shows this
on the participant page:

    Survey       Responses
    ------       ---------
  survey_title   response_set_title
                 response_set_title

There is no way to visually differentiate between these two response
sets aside from the URL that may appear in a status bar or some other
widget.

Our "and I follow" step operates on link title, just like most humans
would, so this issue of Cases' UI affects our test harness as well.

This commit introduces a hack to get around this issue.  In the future,
we should find a user-meaningful way to differentiate between response
sets and eliminate this step.