Build PscParticipants from a Superposition's participants.  #2725.
For those wondering: Yes, we do still need the code that derives
participants from event and instrument data; Field does not return
existing participants in its response.  I suppose it _could_, but this
is so far the only reason to do so, and IMO it doesn't make a very
strong case.