Revert "Update Schema"
This reverts commit bd4a1d9c1ca9aecb85ec7c909f29a5db25587f73.

This commit includes what appears to be the result of running `bundle
update` unqualified. Given the commit message, it is not clear
whether this was actually intended.

Among other things, the new version of Surveyor brought in this way
is not compatible with the actual survey definition files for Cases.
Gems should be updated carefully and mindfully. Newer is not always
better.