If Merge is instructed to not sync and we reach the sync bit, return true.
This was breaking tests.

I'm a bit annoyed that a conditional inserted for the purposes of making
tests pass is causing test failures.  It seems to me that there should
be a better way to do this.  (Having a test instance of PSC would be
best, but that's a ways off.)