Fix content negotiation for fieldwork features.
For those watching: If you need to work with content negotiation in your
own tests, don't do it with extensions unless it's too much of a pain to
set the correct headers for each request.  You should, in my opinion, do
this via the Accept (and Accept-Language) headers, as that's what they
were designed for.

(All major Web browers totally screw up those headers, yes, but that's
irrelevant.  With API clients we can specify tighter expectations on
this sort of thing.)