Spec profiling.
Current CI build times are:

1. _very_ long (an hour and a half!)
2. embarassing

As a result, the CI build's usefulness has dropped dramatically.

Examples that use Surveyor are known to be slow.  However, we don't have
data on which examples those are.  This commit enables RSpec's profiling
to give us some data on which examples to attack first.