Inject Mdes::Specification objects for testing.  #2811.
There's a scenario in api/code_lists.feature that needs to control
the modification time of the MDES disposition code source file.
Unfortunately, the MDES specification used by Cases is a global.

This bit of hackery swaps in a brand-new Mdes::Specification object that
can be modified at will, and then fixes it back up after a scenario run.