Simplify binning. #2864
Bins are per-table_ident also. Move that criterion into
ResponseBin#will_accept? with the others for consistency.

This is a bit awkward because the fact that a table_ident can be
determined is the criterion for attempting to bin a response at all.
I think this code is still clearer than the old version, though.