Make PatientStudyCalendar's Logger configurable.  #2812.
This commit sets all PatientStudyCalendar instances constructed via
ApplicationController#psc to a logger object that writes to log/psc.log,
and defaults all other instances to write to stderr.  Examples for
PatientStudyCalendar use a StringIO object that can be inspected in
examples.

Aside from configurability, this also means that #log no longer writes
to a class-level variable.  This eliminates a source of unsynchronized
runtime modification of shared state, which is important for being able
to use a PatientStudyCalendar instance from multiple threads, which is
what #2812 will end up doing.  (I don't know if this makes
PatientStudyCalendar thread-safe, but it is a step.)