Fieldwork now stores derived entities.  #2812.
Previous to #2812, a fieldwork set was derived entirely from the
PSC scheduled activity report.  Now, however, it's also derived from PSC
schedule previews.  I don't think it makes sense to lump that in the
scheduled activity report.

Hence, this change.  This lumps all the data needed to build the JSON
representation of a fieldwork set into Fieldwork, and makes said data
serializable to boot.  (I think it'll be a good idea for debugging.)