Refine automatic start date determination. #2849.
The previous method assumed that the only contacts without event
start dates would be consecutive contacts for the same event. This
turns out not to be the case in the real data.

This commit will only take the event start date from the previous
contact if the participant and event type are the same. If they are
not but the event start date is missing, it reports an error and then
uses the contact date as the event start date.