Synchronize collection modification.  #2812.
Technically, we don't need to synchronize access to *all* collections,
but this is less code and I have no evidence that contention on these
collections is going to be a big deal.  (The point of running these
requests in parallel is to cut down on PSC communication time, and
synchronizing on collections isn't going to block that.)  If said
contention is a big deal, this can be fixed later.