Separate Monit config generation and installation.  #2894.
This means that the Rake task is no longer run as root.  (On our
systems, it also *can't* run as root, because root does not
automatically use the same Ruby environment as applications.)

The umask bit is important, and should not be changed without a very
good reason.  Without it, the deploy task will still run, but generated
configuration may be viewable and modifiable by another user on the app
server.