Flip wait and worker check.  #2893.
The Timer actor is supervised.  This is a good thing, but in the case of
i.e. a prolonged Redis crash, doing a check-before-wait will lock the
supervisor in a tight restart loop.  On Rubies with GILs (which is what
we're using) this will bring work in other threads to a halt.

(AFAICT, Celluloid's supervision groups do not implement backoff -- they
restart the crashed actor immediately.)