Reserve six columns for PID.
Reserving a fixed number of spaces helps formatting and makes log
streams easier to read.

It's possible to exceed this in some situations: the Linux kernel
currently permits PIDs up to 2^29, for example.  However, most systems
(including all of ours) default to a PID ceiling < 999999, so this is
good enough for now.  If we ever start running Cases on e.g. a Cray XK6
then we'll worry about this then.