date picker submit on blur #2886
StaleObjectError messages are occurring in many Cases instances
These are mostly due to the 6 requests posted from the jquery date picker
launched from the change event.
This commit moves the .hasDatepicker and .datetime bindings to the blur event
while keeping all other change events the same.

I am hesitant to update the jquery.surveyor.js that is copied from the surveyor gem
because this will be updated anytime surveyor is re-installed but this file is already
out of sync with the gem's file so that issue can be addressed at another time.

Fixes #2886