Use Time.parse in Response#value=.  Closes #2963.
Chronic.parse cannot parse ISO8601 datetimes of the form
YYYY-MM-DDTHH:MM+/-HHMM.  However, Surveyor now sends datetimes in that
format, so we need to handle that.

Time.parse works just fine, though.

The ISO8601_HHMM_FORMAT regexp is present to let us know when we should
interpret an incoming string as a datetime.  This is required because
JSON (which is what i.e. Surveyor iOS gives us when we get data
back from Field) does not have a specialized datetime type.