A Huge Hack™ to sort out operational data extraction.
Here's the problem:

1. The response set for a response must have been persisted before the
   response is saved.
2. Operational data extractors run after a response set is saved.
3. Operational data extractors use responses.

If response sets are saved before responses, there is no data for the
extractors to use.  If response sets are saved after responses,
referential integrity is violated.

This Huge Hack™ just saves response sets twice.  (After-save hooks fire
on clean records.)  It's slow and gross but should work.

A Much Better Fix™ involves redesigning the save code to account for
cases like this, but I don't have the time to do that right
now.  Expect to see the Much Better Fix™ in Future Commits™.