Consistency fix: remove private.  #3006.
No other methods in Psc::SyncLoader are declared private except via
documentation tags.

IMO, private methods in Ruby don't buy you very much _except_
documentation.

It's true that private methods don't show up in .methods and that using
an object via a REPL is a great way to understand how it behaves, but I
think that if you want to understand the object's interface you really
need to read its documentation -- and once you do that, you'll find out
what's considered to be private.

As far as I know, we do not have a codebase convention for this.  If
one's developed, I can reinstate this scope for all helper methods then.