A script for experimenting with merges.  #3007.
This is almost done; I still need to find a good way to load up surveys
from the private instruments repo.  But it's close enough that I think
committing it to master is fine.

The merge process involves three datasets and is therefore not trivial
to debug when things go wrong.

On the other hand, the merge behavior is deterministic.  Therefore
things that go wrong -- while they can be hard to understand -- are, in
theory, easy to reproduce *if* you can control those three datasets.

This script is a step towards turing theory into practice.  Ultimately,
it should be used to derive minimized test cases from staging and
production data.