List prepopulated responses #3024
List prepopulated responses which are derivable from context for event templates

This is a file that will be used temporarily by field to prepopulate known values for
initial instruments during the PBS eligibility and PV1 events. These instruments are
the PBS eligibility screener, Tracing Module, Participant Verification, and
Pregnancy Visit 1.

Closes #3024