nil provider for recruitment event #2858
Added check for nil when determining redirect path after
event update. The provider /should/ exist for a provider recruitment event
but it was seen during training that the provider was not found in the event
collection of contact links. In which case we redirect the user to the pbs_lists
index page rather than the specific pbs_list provider page.

Closes #2858