Canceling activities in PSC #2839
The previous commit f564512cd84d1d31e7884edecc1e585fcbf5b1dc
removed the call to cancel activities in PSC when that activity did not have
a matching MDES version. This was done because the PSC template did
not have all the instruments labelled properly and we were canceling activities
that should not have been canceled.

This change reverts that commit since we now have the PSC template updated
to include the properly labelled instruments AND more importantly we are
scheduling 3.0 activities in a 2.0 environment. By re-instating the cancel call
we can move forward with importing data for 2.x study centers.

This commit also includes a rake task to cancel activities that have an instrument
for an MDES version that does not match the current MDES version. This task has been
run in the UMN staging and production environments.