Check for nil #3113
Update to surveyor dependency condition hash to check for
nil before sending call to non-existent method on NilClass.

This might need to be added to Surveyor as this patch is the same
code for to_hash as it is in Surveyor 1.0. (and the comment at the
top of this patch clearly states that this is a temporary fix until we
update the Surveyor version used by Cases).