Lessen events window #2665
This commit is a revert of #2665. The original task
was to include all events during a training. This commit
brings the window of upcoming events back into a more
useful date range for most activities.

Plus these values are defaults and can be updated in most
searches through the UI.