Update activity state in PSC #3097
These changes are not intended to make it into the final
version for either #3097 or #3125. This is a spike attempt at
updating the activities in PSC through the UI similar to the
proposed changes for the event workflow in #3125. The
lack of tests shows the 'spike-ness' of this commit.

However, a demonstration for stakeholders requires some of this
functionality to make it to the intended participant state thus this
commit will be tagged and deployed to a private demo instance of
the application.