Event#disposition_complete? -> Event#completed?.  #3128.
This commit also shortens up the examples for Event#completed?, as they
all follow the same format and fall pretty well into a table-driven
style of testing.

It's not quite as nice as I'd like (Cucumber-style tables would be more
readable), but IMO it isn't _that_ bad.  Further refinements if I think
of them.