Read response set public ID from a Response's source or target.  #3146.
Useful because a response may come from the database or from data sent
by Field.  In the latter case, we have direct access to the response's
ancestry; in the former case, we don't.  However, we still have access
to the response's response set API ID.

With this change, if you instantiate a Response::ModelAdapter around
a Response and then change the adapter's source to some
Response::HashAdapter, then the response set public ID of the
Response::ModelAdapter instance will change.  I think that's okay,
becaues I don't have an example of a situation in which someone would do
that, and this merge code isn't intended to be used as a library.