Add static mappings for person_id and ppg_first. Closes #3191.
Neither of these is ideal, but this fix requires a quick turnaround
and these implementations are good enough for now. Some detailed
problems with ppg_first are listed in #3193.

The problems with person_id are obvious, but making it more response-
set aware would require substantial engineering effort. Since the
implementation does work for all known instruments at this point, I'm
leaving that for another day.