Set defaults for p_info_XXX #3218
These fields are in the person form but can be skipped.
Here we set probable defaults and, of course, allow the
user to edit them. This means the user can always update
these to be blank, but the only required field is the info_source
which if updated to nil will be set to the default of -4 (Missing in Error)
upon saving of the record.

Closes #3218