Introduce Clockwork for scheduling jobs.  #3127.
I wanted to keep scheduling inside Cases' web workers, but that has a
few problems:

1. It requires locking to avoid duplicate scheduling, which wasn't
   implemented.
2. It doesn't actually work in Passenger under MRI 1.9.3.  The way those
   two components interact prevents background threads from running.

Solution: move scheduling into a separate process.

I'd like to keep the scheduler process footprint very small.  This means
that the scheduler code (if I can help it) will *not* run through Cases'
boot process.

Since the scheduler is just triggering jobs to run in a Sidekiq process
this should be doable, but it may require some surgery to separate
configuration code (which, except for the Rails environment, can be done
without any references to the Rails object) from code that uses said
configuration to modify the Rails object.